quick commit

+++ PathFinder/Grid/CoordNeighbors.cs
@@ -4,7 +4,7 @@ namespace Zenseless.PathFinder.Grid
 {
 	public static class CoordNeighbors
 	{
-		public static IEnumerable<Coord> Get4Neighbors(this Coord pos, ushort columns, ushort rows)
+		public static IEnumerable<Coord> Get4Neighbors(this Coord pos, int columns, int rows)
 		{
 			var top = pos.Row < rows - 1;
 			var bottom = 0 < pos.Row;
@@ -16,7 +16,7 @@ namespace Zenseless.PathFinder.Grid
 			if (right) yield return new Coord(pos.Column + 1, pos.Row);
 		}
 
-		public static IEnumerable<Coord> Get8Neighbors(this Coord pos, ushort columns, ushort rows)
+		public static IEnumerable<Coord> Get8Neighbors(this Coord pos, int columns, int rows)
 		{
 			var top = pos.Row < rows - 1;
 			var bottom = 0 < pos.Row;
